requirements.txt
PyMuPDF
Faker
reportlab
Pillow
qrcode
pdfplumber
pytesseract
pdf2image
streamlit
Flask
